requirements.txt
amqp==5.1.1
asgiref==3.6.0
beautifulsoup4==4.11.1
billiard==3.6.4.0
cachetools==5.3.0
celery==5.2.7
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
distlib==0.3.6
Django==4.1.5
django-taggit==3.1.0
embeddify==0.3.1
filelock==3.9.0
google-api-core==2.11.0
google-api-python-client==2.78.0
google-auth==2.16.1
google-auth-httplib2==0.1.0
googleapis-common-protos==1.58.0
gunicorn==20.1.0
httplib2==0.21.0
idna==3.4
insta-scrape==2.1.2
instaloader==4.9.5
kombu==5.2.4
numpy==1.24.1
opencv-python==4.7.0.68
Pillow==9.4.0
platformdirs==2.6.2
prompt-toolkit==3.0.38
protobuf==4.22.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
pytube==12.1.2
pytz==2022.7.1
requests==2.28.2
rsa==4.9
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.14
vine==5.0.0
virtualenv==20.17.1
wcwidth==0.2.6
whitenoise==6.4.0
youtube-dl==2021.12.17
